requirements.txt
torch
ipykernel
pillow
gradio
torchvision
numpy